% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{obsStudyGeneExpressions}
\alias{obsStudyGeneExpressions}
\title{Simulated dataset for package \code{DevTreatRule}}
\format{A data frame with 5000 rows and 15 variables:
\itemize{
  \item no_relapse: Binary indicator of remaining relapse-free at the end of study period
  \item intervention: Binary indicator of receiving treatment
  \item prognosis: Two-level factor variable indicating whether clinic-specific measure of prognosis is ``poor'' or `good''
  \item clinic: Ten-level factor variable indicating which of 10 clinics each observation attended
  \item age: Numeric variable indicating an individual's age (ranges from 40 to 65)
  \item gene_1: Numeric variable representing the expression level of specific gene_1
  \item gene_2: Numeric variable representing the expression level of specific gene_2
  \item gene_3: Numeric variable representing the expression level of specific gene_3
  \item gene_4: Numeric variable representing the expression level of specific gene_4
  \item gene_5: Numeric variable representing the expression level of specific gene_5
  \item gene_6: Numeric variable representing the expression level of specific gene_6
  \item gene_7: Numeric variable representing the expression level of specific gene_7
  \item gene_8: Numeric variable representing the expression level of specific gene_8
  \item gene_9: Numeric variable representing the expression level of specific gene_9
  \item gene_10: Numeric variable representing the expression level of specific gene_10
}}
\usage{
obsStudyGeneExpressions
}
\description{
Simulated observational dataset reporting an indicator of remaining relapse-free at the end of the study period (clinical outcome) along with 14 additional baseline characteristics that may influence either treatment assignment, response to treatment, or both.
}
\keyword{datasets}
